rearrange health checks to clarify what the PrometheusHealthResponseWrapper does
